Remove hpux, irix, sunos support. Use floating point audio writing. Temporarily break most audio drivers, Linux/Pulse still works.
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-libs-2.1.1@25 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/resid-builder/src/resid.cpp
@@ -64,7 +64,7 @@ ReSID::ReSID (sidbuilder *builder)
         return;
     }
 
-    m_buffer = new short[OUTPUTBUFFERSIZE];
+    m_buffer = new float[OUTPUTBUFFERSIZE];
     m_bufferpos = 0;
     reset (0);
 }
@@ -133,9 +133,9 @@ void ReSID::clock()
     cycle_count cycles = m_context->getTime(m_accessClk, m_phase);
     m_accessClk += cycles;
     if (m_optimisation)
-        m_bufferpos += m_sid.clock_fast(cycles, (short *) m_buffer + m_bufferpos, OUTPUTBUFFERSIZE - m_bufferpos, 1);
+        m_bufferpos += m_sid.clock_fast(cycles, (float *) m_buffer + m_bufferpos, OUTPUTBUFFERSIZE - m_bufferpos, 1);
     else
-        m_bufferpos += m_sid.clock(cycles, (short *) m_buffer + m_bufferpos, OUTPUTBUFFERSIZE - m_bufferpos, 1);
+        m_bufferpos += m_sid.clock(cycles, (float *) m_buffer + m_bufferpos, OUTPUTBUFFERSIZE - m_bufferpos, 1);
 }
 
 void ReSID::filter (bool enable)
